version 19 09 2026 Prédéclaration des contraintes business.

+++ EconomicTrendsPolLESSIRE.Application/Interfaces/IInstrumentService.cs
@@ -5,11 +5,102 @@ namespace EconomicTrendsPolLESSIRE.Application.Interfaces
 {
     public interface IInstrumentService
     {
-        Task<InstrumentDTO?> GetByIdAsync(int id);
+        Task<InstrumentDTO?> GetByIdAsync(long id);
         Task<InstrumentDTO?> SaveAsync(InstrumentDTO dto);
         Task<EconomicTrendsPolLESSIRE.Domain.Entities.Instrument?> SaveInstrumentAsync(EconomicTrendsPolLESSIRE.Domain.Entities.Instrument instrument, CancellationToken ct = default);
         Task<IEnumerable<EconomicTrendsPolLESSIRE.Domain.Entities.Instrument>> GetAllAsync(int limit = 500, CancellationToken ct = default);
-        Task<EconomicTrendsPolLESSIRE.Domain.Entities.Instrument?> GetInstrumentByIdAsync(int id, CancellationToken ct = default);
-        Task<bool> DeleteInstrumentAsync(int id, CancellationToken ct = default);
+        Task<EconomicTrendsPolLESSIRE.Domain.Entities.Instrument?> GetInstrumentByIdAsync(long id, CancellationToken ct = default);
+        Task<bool> DeleteInstrumentAsync(long id, CancellationToken ct = default);
     }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// Copyrigtht (c) EconomicTrendsPolLESSIRE https://github.com/POLLESSI/EconomicTrendsPolLESSIRE. All rights reserved.
