+++ EconomicTrendsPolLESSIRE.Domain/Entities/MarketQuote.cs
@@ -3,7 +3,7 @@
     public class MarketQuote
     {
         public int Id { get; set; }
-        public int InstrumentId { get; set; }
+        public long InstrumentId { get; set; }
         public int ProviderId { get; set; }
         public DateTime TimestampUtc { get; set; }
         public DateTime ReceivedAtUtc { get; set; }
@@ -14,3 +14,85 @@
         public bool Active { get; set; }
     }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// Copyrigtht (c) EconomicTrendsPolLESSIRE https://github.com/POLLESSI/EconomicTrendsPolLESSIRE. All rights reserved.
