+++ EconomicTrendsPolLESSIRE.Domain/Entities/MarketSnapshot.cs
@@ -2,7 +2,7 @@
 {
     public class MarketSnapshot
     {
-        public int InstrumentId { get; set; }
+        public long InstrumentId { get; set; }
         public decimal LastPrice { get; set; } = 0;
         public decimal BidPrice { get; set; } = 0;
         public decimal AskPrice { get; set; } 
@@ -17,3 +17,97 @@
         public bool Active { get; set; }
     }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// Copyrigtht (c) EconomicTrendsPolLESSIRE https://github.com/POLLESSI/EconomicTrendsPolLESSIRE. All rights reserved.
