+++ EconomicTrendsPolLESSIRE.Domain/Entities/ProviderInstrument.cs
@@ -4,10 +4,100 @@
     {
     #nullable disable
         public int ProviderId { get; set; }
-        public int InstrumentId { get; set; }
+        public long InstrumentId { get; set; }
         public string ProviderSymbol { get; set; }
         public bool Realtime { get; set; }
         public int DelaySeconds { get; set; }
         public bool Active { get; set; }
     }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// Copyrigtht (c) EconomicTrendsPolLESSIRE https://github.com/POLLESSI/EconomicTrendsPolLESSIRE. All rights reserved.
