+++ EconomicTrendsPolLESSIRE.Domain/Entities/TechnicalIndicator.cs
@@ -2,7 +2,7 @@
 {
     public class TechnicalIndicator
     {
-        public int InstrumentId { get; set; }
+        public long InstrumentId { get; set; }
         public int IntervalCode { get; set; }
         public DateTime TimestampUtc { get; set; }
         public int IndicatorType { get; set; }
@@ -13,3 +13,87 @@
         public bool Active { get; set; }
     }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// Copyrigtht (c) EconomicTrendsPolLESSIRE https://github.com/POLLESSI/EconomicTrendsPolLESSIRE. All rights reserved.
